requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
python-decouple==3.6
pytz==2022.6
requests==2.28.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
